Counts are unsigned

+++ src/main.rs
@@ -227,7 +227,7 @@ fn fetch_comments(url: hyper::Uri) -> impl Future<Item=(Vec<Comment>, String), E
 struct Story {
     title: String,
     created_at: String,
-    score: i16,
+    score: u8,
     comment_count: u8,
     short_id: String,
     short_id_url: String,
@@ -251,7 +251,7 @@ struct CommentRoot {
 struct Comment {
     comment: String,
     created_at: String,
-    score: i16,
+    score: u8,
     indent_level: u8,
     commenting_user: User,
 }
